refactor(gui): split workspace command adapters
